refactor: lotto number 패키지 구조화
