+++ src/main/java/model/LottoBonusNumber.java
@@ -1,5 +1,7 @@
 package model;
 
+import model.lottonumber.LottoNumber;
+
 public record LottoBonusNumber(
         LottoNumber number
 ) {
