+++ src/main/java/model/LottoWinningGame.java
@@ -1,5 +1,7 @@
 package model;
 
+import model.lottonumber.LottoNumber;
+
 public record LottoWinningGame(
         LottoGame game
 ) {
