+++ src/test/java/model/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.lottonumber.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
