+++ src/test/java/model/LottoTicketSellerTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.lottonumber.LottoNumberGeneratorImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
