+++ src/test/java/model/LottoWinningNumbersTest.java
@@ -1,5 +1,7 @@
 package model;
 
+import model.lottonumber.FakeLottoNumberGenerator;
+import model.lottonumber.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
